# Improve performance of urldecode() and rawurldecode()